# web/openapi.json
{
  "components": {
    "schemas": {
      "AllocationRow": {
        "properties": {
          "end_price": {
            "title": "End Price",
            "type": "number"
          },
          "return_pct": {
            "title": "Return Pct",
            "type": "number"
          },
          "start_price": {
            "title": "Start Price",
            "type": "number"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "ticker",
          "weight",
          "start_price",
          "end_price",
          "return_pct"
        ],
        "title": "AllocationRow",
        "type": "object"
      },
...
      "BacktestMetrics": {
        "properties": {
          "alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alpha"
          },
          "annualized_return_pct": {
            "title": "Annualized Return Pct",
            "type": "number"
          },
          "ath": {
            "title": "Ath",
            "type": "number"
          },
          "atl": {
            "title": "Atl",
            "type": "number"
          },
          "beta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "max_drawdown": {
            "title": "Max Drawdown",
            "type": "number"
          },
          "optimal_sharpe": {
            "title": "Optimal Sharpe",
            "type": "number"
          },
          "sharpe_ratio": {
            "title": "Sharpe Ratio",
            "type": "number"
          },
          "total_return_pct": {
            "title": "Total Return Pct",
            "type": "number"
          }
        },
        "required": [
          "ath",
          "atl",
          "total_return_pct",
          "annualized_return_pct",
          "max_drawdown",
          "sharpe_ratio",
          "optimal_sharpe"
        ],
        "title": "BacktestMetrics",
        "type": "object"
      },
      "BacktestRequest": {
        "properties": {
          "initial_capital": {
            "default": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Initial Capital",
            "type": "number"
          },
          "max_share": {
            "default": 0.4,
            "description": "Maximum weight for any single asset",
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Max Share",
            "type": "number"
          },
          "min_share": {
            "default": 0.0,
            "description": "Minimum weight threshold; assets below are dropped",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Min Share",
            "type": "number"
          },
          "opt_end": {
            "description": "End of optimisation window and start of investment",
            "format": "date",
            "title": "Opt End",
            "type": "string"
          },
          "opt_start": {
            "description": "Start of look-back / optimisation window",
            "format": "date",
            "title": "Opt Start",
            "type": "string"
          },
          "rebalance_freq": {
            "default": 0,
            "description": "Rebalance every N business days. 0 = static allocation",
            "minimum": 0.0,
            "title": "Rebalance Freq",
            "type": "integer"
          },
          "tickers": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "minItems": 1,
            "title": "Tickers",
            "type": "array"
          }
        },
        "required": [
          "tickers",
          "opt_start",
          "opt_end"
        ],
        "title": "BacktestRequest",
        "type": "object"
      },
      "BacktestResponse": {
        "properties": {
          "allocation": {
            "items": {
              "$ref": "#/components/schemas/AllocationRow"
            },
            "title": "Allocation",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/BacktestMetrics"
          },
          "opt_end": {
            "title": "Opt End",
            "type": "string"
          },
          "opt_start": {
            "title": "Opt Start",
            "type": "string"
          },
          "portfolio_values": {
            "items": {
              "$ref": "#/components/schemas/PortfolioPoint"
            },
            "title": "Portfolio Values",
            "type": "array"
          },
          "sp500_values": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PortfolioPoint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sp500 Values"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "opt_start",
          "opt_end",
          "allocation",
          "portfolio_values",
          "metrics",
          "warnings"
        ],
        "title": "BacktestResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortfolioPoint": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "date",
          "value"
        ],
        "title": "PortfolioPoint",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "quantModeling API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/backtest/run": {
      "post": {
        "description": "Optimise a portfolio of assets over a historical window, then track its\nperformance forward. Returns allocation, portfolio value time series,\nperformance metrics, and CAPM alpha/beta vs S&P 500.",
        "operationId": "run_backtest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Backtest",
        "tags": [
          "backtest"
        ]
      }
    },
...
  }
}